feat: Make cache policy compatible wit V4 API
BREAKING CHANGE: This implementation is using the dependencies introduced by Gravitee V4.0
