+++ README.adoc
@@ -34,6 +34,7 @@ policy configuration `cacheName`, as described below.
 | Plugin version | APIM version
 
 | Up to 1.x                   | All
+| From 2.x                   | 4.0+
 |===
 
 == Configuration
