+++ src/main/java/io/gravitee/policy/v3/cache/proxy/CacheProxyConnection.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.policy.cache.proxy;
+package io.gravitee.policy.v3.cache.proxy;
 
 import io.gravitee.gateway.api.buffer.Buffer;
 import io.gravitee.gateway.api.handler.Handler;
