+++ src/main/java/io/gravitee/policy/v3/cache/proxy/EvaluableProxyResponse.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.policy.cache.proxy;
+package io.gravitee.policy.v3.cache.proxy;
 
 import io.gravitee.gateway.api.Response;
 import io.gravitee.gateway.api.el.EvaluableResponse;
