+++ src/main/resources/plugin.properties
@@ -6,3 +6,4 @@ class=io.gravitee.policy.cache.CachePolicy
 type=policy
 category=performance
 icon=cache.svg
+proxy=REQUEST
