+++ src/test/resources/io/gravitee/policy/cache/integration/cacheV3.json
